feat(database): duplication and code refactoring
