+++ modules/database/src/adapters/sequelize-adapter/postgres-adapter/PostgresSchema.ts
@@ -1,9 +1,9 @@
-import { isNil } from 'lodash';
 import { ModelStatic, Sequelize, Transaction } from 'sequelize';
-import { ParsedQuery, SingleDocQuery } from '../../../interfaces';
+import { SingleDocQuery } from '../../../interfaces';
 import ConduitGrpcSdk, { ConduitSchema, Indexable } from '@conduitplatform/grpc-sdk';
 import { SequelizeSchema } from '../SequelizeSchema';
 import { PostgresAdapter } from './index';
+import { getTransactionAndParsedQuery, processPushOperations } from '../utils';
 
 const incrementDbQueries = () =>
   ConduitGrpcSdk.Metrics?.increment('database_queries_total');
@@ -29,20 +29,11 @@ export class PostgresSchema extends SequelizeSchema {
     populate?: string[],
     transaction?: Transaction,
   ): Promise<{ [key: string]: any }> {
-    let t: Transaction | undefined = transaction;
-    const transactionProvided = transaction !== undefined;
-    let parsedQuery: ParsedQuery;
-    if (typeof query === 'string') {
-      parsedQuery = JSON.parse(query);
-    } else {
-      parsedQuery = query;
-    }
-    if (parsedQuery.hasOwnProperty('$set')) {
-      parsedQuery = parsedQuery['$set'];
-    }
-    if (isNil(t)) {
-      t = await this.sequelize.transaction({ type: Transaction.TYPES.IMMEDIATE });
-    }
+    const { t, parsedQuery, transactionProvided } = await getTransactionAndParsedQuery(
+      transaction,
+      query,
+      this.sequelize,
+    );
     try {
       const parentDoc = await this.model.findByPk(id, {
         nest: true,
@@ -67,29 +58,7 @@ export class PostgresSchema extends SequelizeSchema {
 
       if (parsedQuery.hasOwnProperty('$push')) {
         const push = parsedQuery['$push'];
-        for (const key in push) {
-          if (this.extractedRelations[key]) {
-            if (!Array.isArray(this.extractedRelations[key])) {
-              throw new Error(`Cannot push in non-array field: ${key}`);
-            }
-            let modelName = key.charAt(0).toUpperCase() + key.slice(1);
-            if (push[key]['$each']) {
-              if (!modelName.endsWith('s')) {
-                modelName = modelName + 's';
-              }
-              parentDoc[`add${modelName}`](push[key]['$each'], parentDoc._id);
-            } else {
-              const actualRel = key.charAt(0).toUpperCase() + key.slice(1);
-              parentDoc[`add${actualRel}Id`](push[key], parentDoc._id);
-            }
-            continue;
-          }
-          if (push[key]['$each']) {
-            parentDoc[key] = [...parentDoc[key], ...push[key]['$each']];
-          } else {
-            parentDoc[key] = [...parentDoc[key], push[key]];
-          }
-        }
+        await processPushOperations(parentDoc, push, this.extractedRelations);
         await parentDoc.save();
         delete parsedQuery['$push'];
       }
