+++ modules/database/src/adapters/sequelize-adapter/sql-adapter/index.ts
@@ -1,11 +1,10 @@
 import { SQLSchema } from './SQLSchema';
 import { schemaConverter } from './SchemaConverter';
-import { ConduitSchema, sleep } from '@conduitplatform/grpc-sdk';
-import { validateFieldChanges, validateFieldConstraints } from '../../utils';
+import { ConduitSchema } from '@conduitplatform/grpc-sdk';
 import { isNil } from 'lodash';
 import { ConduitDatabaseSchema } from '../../../interfaces';
 import { SequelizeAdapter } from '../index';
-import { SequelizeSchema } from '../SequelizeSchema';
+import { compileSchema, registerAndResolveRelatedSchemas } from '../utils';
 
 export class SQLAdapter extends SequelizeAdapter<SQLSchema> {
   constructor(connectionUri: string) {
@@ -64,74 +63,20 @@ export class SQLAdapter extends SequelizeAdapter<SQLSchema> {
     schema: ConduitSchema,
     options?: { parentSchema: string },
   ): Promise<SQLSchema> {
-    let compiledSchema = JSON.parse(JSON.stringify(schema));
-    validateFieldConstraints(compiledSchema);
-    (compiledSchema as any).fields = JSON.parse(
-      JSON.stringify((schema as ConduitDatabaseSchema).compiledFields),
+    const compiledSchema = compileSchema(
+      schema as ConduitDatabaseSchema,
+      this.registeredSchemas,
+      this.sequelize.models,
     );
-    if (this.registeredSchemas.has(compiledSchema.name)) {
-      if (compiledSchema.name !== 'Config') {
-        compiledSchema = validateFieldChanges(
-          this.registeredSchemas.get(compiledSchema.name)!,
-          compiledSchema,
-        );
-      }
-      delete this.sequelize.models[compiledSchema.collectionName];
-    }
 
     const [newSchema, extractedSchemas, extractedRelations] =
       schemaConverter(compiledSchema);
-    this.registeredSchemas.set(
-      schema.name,
-      Object.freeze(JSON.parse(JSON.stringify(schema))),
+    const relatedSchemas = await registerAndResolveRelatedSchemas(
+      schema as ConduitDatabaseSchema,
+      this.registeredSchemas,
+      extractedRelations,
+      this.models,
     );
-    const relatedSchemas: { [key: string]: SequelizeSchema | SequelizeSchema[] } = {};
-
-    if (Object.keys(extractedRelations).length > 0) {
-      let pendingModels: string[] = [];
-      for (const relation in extractedRelations) {
-        const rel = Array.isArray(extractedRelations[relation])
-          ? (extractedRelations[relation] as any[])[0]
-          : extractedRelations[relation];
-        if (!this.models[rel.model] || !this.models[rel.model].synced) {
-          if (!pendingModels.includes(rel.model)) {
-            pendingModels.push(rel.model);
-          }
-          if (Array.isArray(extractedRelations[relation])) {
-            relatedSchemas[relation] = [rel.model];
-          } else {
-            relatedSchemas[relation] = rel.model;
-          }
-        } else {
-          if (Array.isArray(extractedRelations[relation])) {
-            relatedSchemas[relation] = [this.models[rel.model]];
-          } else {
-            relatedSchemas[relation] = this.models[rel.model];
-          }
-        }
-      }
-      while (pendingModels.length > 0) {
-        await sleep(500);
-        pendingModels = pendingModels.filter(model => {
-          if (!this.models[model] || !this.models[model].synced) {
-            return true;
-          } else {
-            for (const schema in relatedSchemas) {
-              // @ts-ignore
-              const simple = Array.isArray(relatedSchemas[schema])
-                ? (relatedSchemas[schema] as SequelizeSchema[])[0]
-                : relatedSchemas[schema];
-              // @ts-ignore
-              if (simple === model) {
-                relatedSchemas[schema] = Array.isArray(relatedSchemas[schema])
-                  ? [this.models[model]]
-                  : this.models[model];
-              }
-            }
-          }
-        });
-      }
-    }
     const associatedSchemas: { [key: string]: SQLSchema | SQLSchema[] } = {};
     await this.processExtractedSchemas(schema, extractedSchemas, associatedSchemas);
     if (options?.parentSchema) {
