+++ modules/database/src/controllers/customEndpoints/utils.ts
@@ -6,11 +6,7 @@ import ConduitGrpcSdk, {
 } from '@conduitplatform/grpc-sdk';
 import { ICustomEndpoint, PopulatedCustomEndpoint } from '../../interfaces';
 import { isNil } from 'lodash';
-export const LocationEnum = {
-  BODY: 0,
-  QUERY: 1,
-  URL: 2,
-};
+import { LocationEnum } from '../../enums';
 
 function getOperation(op: number) {
   switch (op) {
