+++ modules/database/src/enums/index.ts
@@ -1,0 +1,2 @@
+export * from './operation.enum';
+export * from './location.enum';
