+++ modules/database/src/enums/location.enum.ts
@@ -1,0 +1,5 @@
+export enum LocationEnum {
+  BODY = 0,
+  QUERY = 1,
+  URL = 2,
+}
