+++ modules/database/src/enums/operation.enum.ts
@@ -1,0 +1,7 @@
+export enum OperationsEnum {
+  GET = 0,
+  POST = 1,
+  PUT = 2,
+  DELETE = 3,
+  PATCH = 4,
+}
